Fix Win32 exit exit val tests
Shell quoting can't be done on `cmd.exe` or its options, and the script must
be passed as a single item. So do the quoting manually. This highlights that
cmd.exe calls may be broken compared to how they were before. If we wanted to
fix that, it would probably require examining the $exe and declining to shell
quote args if it's cmd.exe or command.exe. I can't say whether that's
desireable.

Also, fix run() so that it uses Win32::Process::Create() instead of system and
properly checks the exit values. This gets all Win32 tests passing except for
RT#48319.